Extract code into new component
