+++ src/components/CountrySelect/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "CountrySelect.js",
+  "version": "0.0.0",
+  "main": "./CountrySelect.js"
+}
